Add better error handling for server error 500.
Some users were running into issues where the filename was breaking
the upload. This should no longer happen for the main cases,
but it was a quick fix so there may still be edge cases.
The error is now shown on the page rather than an unhelpful
"Server Error 500" page.
Also remove linode object storage upload because it seems we don't
need it.